objtype: include the real objGetNearestTypeTo prototype in the 21 implicit-declaration callers
Every TU that called objGetNearestTypeTo through an implicit int
declaration now sees the prototype from main/objtype.h. Redundant
(GameObject*) result casts dropped; int-receiving sites (294 t/t2,
529 wmwallcrawler_update player, 644 vendorObj) spell the pointer-to-int
conversion explicitly. All 21 objects certified neutral in every linked
section, relocation and symbol; main.dol byte-identical.

+++ src/dlls/objects/644/644.c
@@ -23,6 +23,7 @@
 #include "main/dll/player_api.h"
 #include "main/curve_eval.h"
 #include "main/objseq.h"
+#include "main/objtype.h"
 #include "main/dll/dll_0284_shopitem.h"
 #include "main/dll/LGT/LGTcontrollight.h"
 #include "main/dll/boulder.h"
@@ -313,7 +314,7 @@ void shopitem_update(GameObject* obj)
         if (*(u32*)&s->vendorObj == 0)
         {
             int item;
-            s->vendorObj = objGetNearestTypeTo(SHOPITEM_TARGET_OBJGROUP, obj, &range);
+            s->vendorObj = (int)objGetNearestTypeTo(SHOPITEM_TARGET_OBJGROUP, obj, &range);
             item = s->vendorObj;
             if ((u32)item != 0)
             {
